requirements.txt
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.0
bs4==0.0.1
cffi==1.14.0
click==7.1.1
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
Flask==1.1.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
numpy==1.18.4
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-slugify==4.0.0
queuelib==1.5.0
Scrapy==2.1.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
text-unidecode==1.3
Twisted @ file:///C:/Users/Ondra/Documents/Skola/pageskunk/venv/Twisted-20.3.0-cp38-cp38-win_amd64.whl
w3lib==1.21.0
Werkzeug==1.0.0
Whoosh==2.7.4
WTForms==2.3.1
zope.interface==5.0.1
